refactor(mempool): move to `chain_index` mod
